requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
scipy
statsmodels



openai>=1.0.0
python-dotenv==1.0.0
anthropic>=0.10.0
pandas>=2.0.0
tabulate>=0.9.0

# deepseek
torch==2.4.1
triton==3.0.0
transformers==4.46.3
safetensors==0.4.5

    